% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtextrankr.R
\name{build_graph}
\alias{build_graph}
\title{Make a graph which refers relationship of sentences.
Vertex refers one sentence, and edge refers co-occurence between two sentences.
Example will be shown in \href{https://github.com/mikigom/Rtextrankr}{github}.}
\usage{
build_graph(sentences)
}
\arguments{
\item{sentences}{The list of character string.}
}
\value{
The igraph graph which refers relationship of sentences.
}
\description{
Make a graph which refers relationship of sentences.
Vertex refers one sentence, and edge refers co-occurence between two sentences.
Example will be shown in \href{https://github.com/mikigom/Rtextrankr}{github}.
}

