% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtextrankr.R
\name{get_sentences}
\alias{get_sentences}
\title{Make character string list which ends with a period, from the whole text.
Example will be shown in \href{https://github.com/mikigom/Rtextrankr}{github}.}
\usage{
get_sentences(text)
}
\arguments{
\item{text}{The long character string including delimiter([.|!|?|blank line]).}
}
\value{
The list of character string which ends with a period.
}
\description{
Make character string list which ends with a period, from the whole text.
Example will be shown in \href{https://github.com/mikigom/Rtextrankr}{github}.
}

