% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtextrankr.R
\name{sentence2table}
\alias{sentence2table}
\title{Convert sentence to noun word table except stopword.
Example will be shown in \href{https://github.com/mikigom/Rtextrankr}{github}.}
\usage{
sentence2table(sentence)
}
\arguments{
\item{sentence}{The character string.}
}
\value{
A table which refers noun word frequency, except Korean stopword.
}
\description{
Convert sentence to noun word table except stopword.
Example will be shown in \href{https://github.com/mikigom/Rtextrankr}{github}.
}

