% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtextrankr.R
\name{summarize}
\alias{summarize}
\title{Return list of character string of the most important sentences by Textrank algorithm.
Example will be shown in \href{https://github.com/mikigom/Rtextrankr}{github}.}
\usage{
summarize(text, count)
}
\arguments{
\item{text}{The long character string including delimiter([.|!|?|blank line]).}

\item{count}{The number of summarized sentences.}
}
\value{
The list of character string of the most important sentences by Textrank algorithm.
}
\description{
Return list of character string of the most important sentences by Textrank algorithm.
Example will be shown in \href{https://github.com/mikigom/Rtextrankr}{github}.
}
\references{
Mihalcea, R., & Tarau, P. (2004, July). TextRank: Bringing order into texts. Association for Computational Linguistics.
}

