% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtextrankr.R
\name{table2gset}
\alias{table2gset}
\title{Convert sentence to word table except stopword.
Example will be shown in \href{https://github.com/mikigom/Rtextrankr}{github}.}
\usage{
table2gset(table)
}
\arguments{
\item{table}{A table which refers word frequency.}
}
\value{
A generalized set which refers word frequency.
}
\description{
Convert sentence to word table except stopword.
Example will be shown in \href{https://github.com/mikigom/Rtextrankr}{github}.
}

