% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtextrankr.R
\name{xplit}
\alias{xplit}
\title{Separate text into sentence list by delimiter.
Example will be shown in \href{https://github.com/mikigom/Rtextrankr}{github}.}
\usage{
xplit(text)
}
\arguments{
\item{text}{The long character string including delimiter([.|!|?|blank line]).}
}
\value{
The list of character string seperated by delimiter.
}
\description{
Separate text into sentence list by delimiter.
Example will be shown in \href{https://github.com/mikigom/Rtextrankr}{github}.
}

